use std::vec::Vec for headers_row
otherwise headers are resorted and data and header may no longer be
aligned correctly.

e.g. when we have `{"a" :[1,2,3], "a":{"b":2}}` and the flattener uses any
symbol do denote the array path which is sorted after dot (like the
symbol `[`) then usign a BTree for the headers_row potentially results
in resorting.